requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alpaca-py==0.34.0
alpaca-trade-api==3.2.0
annotated-types==0.7.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
Backtesting==0.3.3
beautifulsoup4==4.12.3
bokeh==2.4.3
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
deprecation==2.1.0
executing==2.1.0
finvizfinance==1.1.0
fonttools==4.55.2
frozendict==2.4.6
frozenlist==1.5.0
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mplfinance==0.12.10b0
msgpack==1.0.3
multidict==6.1.0
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
parso==0.8.4
peewee==3.17.8
pendulum==3.0.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
requests==2.32.3
setuptools==75.6.0
six==1.16.0
soupsieve==2.6
sseclient-py==1.8.0
stack-data==0.6.3
ta-lib==0.5.1
time-machine==2.16.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vaderSentiment==3.3.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
xyzservices==2024.9.0
yarl==1.18.3
yfinance==0.2.37
